backend/requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<4.0
Pillow>=8.0,<9.0
PyJWT>=2.0,<3.0
django-cors-headers>=3.7,<4.0
django-filter>=2.4,<3.0